// src/app/shared/generated/model/applicationParameterHistoryCriteria.ts
/**
 * onecx-parameters-bff
 * OneCx parameters Bff
 *
 * The version of the OpenAPI document: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ApplicationParameterHistoryCriteria { 
    applicationId?: string;
    productName?: string;
    key?: string;
    pageNumber?: number;
    pageSize?: number;
    type?: Array<string>;
}