// src/app/shared/generated/model/parameterHistoryCount.ts
/**
 * onecx-parameters-bff
 * OneCx parameters Bff
 *
 * The version of the OpenAPI document: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ParameterHistoryCount { 
    creationDate?: string;
    count?: number;
}